renamed distribution to discrete_distribution
