+++ python_prototype/qmcpy/__init__.py
@@ -8,8 +8,8 @@ from .integrate import integrate
 from ._util import DistributionCompatibilityError, univ_repr, summarize
 from .accum_data.mean_var_data import MeanVarData
 from .accum_data.mean_var_data_rep import MeanVarDataRep
-from .distribution.iid_distribution import IIDDistribution
-from .distribution.quasi_random import QuasiRandom
+from .discrete_distribution.iid_distribution import IIDDistribution
+from .discrete_distribution.quasi_random import QuasiRandom
 from .integrand.asian_call import AsianCall
 from .integrand.keister import Keister
 from .integrand.linear import Linear
