+++ python_prototype/workouts/wo_asian_option.py
@@ -10,7 +10,7 @@ from numpy import arange
 from qmcpy import integrate
 from qmcpy._util import summarize
 from qmcpy.stop import CLT, CLTRep
-from qmcpy.distribution import IIDDistribution, QuasiRandom
+from qmcpy.discrete_distribution import IIDDistribution, QuasiRandom
 from qmcpy.integrand import AsianCall
 from qmcpy.measures import *
